Start forwarding requests from php to node
