+++ twake/backend/core/src/Twake/Discussion/Resources/Services.php
@@ -7,7 +7,8 @@ use Common\BaseServices;
 class Services extends BaseServices
 {
     protected $services = [
-        "app.messages" => "MessageSystem"
+        "app.messages" => "MessageSystem",
+        "app.messages.depreciated" => "MessageSystemDepreciated"
     ];
 
 }
