+++ twake/backend/node/src/services/messages/services/messages/service.ts
@@ -41,7 +41,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
     options?: SaveOptions,
     context?: ThreadExecutionContext,
   ): Promise<SaveResult<Message>> {
-    if (!context?.serverRequest && !this.service.threads.checkAccessToThread(context)) {
+    if (!context?.user?.server_request && !this.service.threads.checkAccessToThread(context)) {
       logger.error(`Unable to write in thread ${context.thread.id}`);
       throw Error("Can't write this message.");
     }
@@ -49,13 +49,15 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
     let message = getInstance({
       id: undefined,
       ephemeral:
-        (context?.app?.id || context?.serverRequest) && item.ephemeral ? item.ephemeral : null,
+        (context?.user?.application_id || context?.user?.server_request) && item.ephemeral
+          ? item.ephemeral
+          : null,
       thread_id: context.thread.id,
-      type: context?.serverRequest && item.type === "event" ? "event" : "message",
+      type: context?.user?.server_request && item.type === "event" ? "event" : "message",
       subtype: getSubtype(item, context),
       created_at: new Date().getTime(),
       user_id: context.user.id,
-      application_id: context?.app?.id || null,
+      application_id: context?.user?.application_id || null,
       text: item.text || "",
       blocks: item.blocks || [],
       files: item.files || null,
@@ -70,7 +72,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
       reactions: null, // Reactions cannot be set on creation
       bookmarks: null,
       override:
-        (context?.app?.id || context?.serverRequest) && item.override
+        (context?.user?.application_id || context?.user?.server_request) && item.override
           ? {
               title: item.override.title,
               picture: item.override.picture,
@@ -92,8 +94,8 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
       // Deleted messages cannot be edited
       if (
         !messageToUpdate ||
-        (!context?.serverRequest &&
-          context?.app?.id !== messageToUpdate.application_id &&
+        (!context?.user?.server_request &&
+          context?.user?.application_id !== messageToUpdate.application_id &&
           context.user.id !== messageToUpdate.user_id) ||
         messageToUpdate.subtype === "deleted"
       ) {
@@ -106,7 +108,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
       };
 
       message = _.assign(messageToUpdate, _.pick(message, "text", "blocks", "files", "context"));
-      if (context?.app?.id || context?.serverRequest) {
+      if (context?.user?.application_id || context?.user?.server_request) {
         message = _.assign(message, _.pick(message, "override"));
       }
     }
@@ -131,7 +133,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
     options: {},
     context: ThreadExecutionContext,
   ): Promise<SaveResult<Message>> {
-    if (!context?.serverRequest && !this.service.threads.checkAccessToThread(context)) {
+    if (!context?.user?.server_request && !this.service.threads.checkAccessToThread(context)) {
       logger.error(`Unable to write in thread ${context.thread.id}`);
       throw Error("Can't edit this message.");
     }
@@ -168,7 +170,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
     options: {},
     context: ThreadExecutionContext,
   ): Promise<SaveResult<Message>> {
-    if (!context?.serverRequest && !this.service.threads.checkAccessToThread(context)) {
+    if (!context?.user?.server_request && !this.service.threads.checkAccessToThread(context)) {
       logger.error(`Unable to write in thread ${context.thread.id}`);
       throw Error("Can't edit this message.");
     }
@@ -235,7 +237,7 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
   }
 
   async delete(pk: Message, context?: ThreadExecutionContext): Promise<DeleteResult<Message>> {
-    if (!context?.serverRequest && !this.service.threads.checkAccessToThread(context)) {
+    if (!context?.user?.server_request && !this.service.threads.checkAccessToThread(context)) {
       logger.error(`Unable to write in thread ${context.thread.id}`);
       throw Error("Can't edit this message.");
     }
@@ -251,9 +253,9 @@ export class ThreadMessagesService implements MessageThreadMessagesServiceAPI {
     }
 
     if (
-      !context?.serverRequest &&
+      !context?.user?.server_request &&
       message.user_id !== context.user.id &&
-      message.application_id !== context?.app?.id
+      message.application_id !== context?.user?.application_id
     ) {
       logger.error(`You have no right to delete this message`);
       throw Error("Can't delete this message.");
@@ -318,11 +320,11 @@ function getSubtype(
   context?: ThreadExecutionContext,
 ): null | "application" | "deleted" | "system" {
   //Application request
-  if (context?.app?.id) {
+  if (context?.user?.application_id) {
     return item.subtype === "application" ? "application" : null;
   }
   //System request
-  else if (context?.serverRequest) {
+  else if (context?.user?.server_request) {
     return item.subtype;
   }
